Add .gitignore file
The node_modules folder contains dependencies that should end up in
the git repository. Likewise, the subdirectories of the cordova folder
are created during cordova build and do not belong to the source.

We can't ignore the entire cordova folder as the config.xml and custom
resources *are* source files and should be checked into the it repo.